refactor: remove nargo
