+++ yarn-project/Dockerfile
@@ -7,12 +7,7 @@ FROM 278380418400.dkr.ecr.eu-west-2.amazonaws.com/yarn-project-base as builder
 COPY . .
 
 # Generate Noir contract TypeScript artifacts.
-COPY --from=noir /usr/src/yarn-project/noir-contracts/target /usr/src/yarn-project/noir-contracts/target
-# Run yarn build to have the json artifacts available for the types generator, generate types, build again.
-RUN apk add perl
-RUN cd /usr/src/yarn-project/noir-contracts && yarn build && ./scripts/types_all.sh && yarn build
-# Cleanup to reduce final image size.
-RUN rm -rf noir-contracts/target
+COPY --from=noir /usr/src/yarn-project/noir-contracts /usr/src/yarn-project/noir-contracts
 
 # Build the entire project.
 RUN yarn tsc -b
