+++ yarn-project/noir-compiler/.dockerignore
@@ -1,0 +1,2 @@
+dest
+node_modules
