+++ yarn-project/noir-compiler/src/index.test.ts
@@ -2,24 +2,9 @@ import { ContractArtifact } from '@aztec/foundation/abi';
 import { LogFn, createDebugLogger } from '@aztec/foundation/log';
 import { fileURLToPath } from '@aztec/foundation/url';
 
-import { execSync } from 'child_process';
 import path from 'path';
 
-import {
-  compileUsingNargo,
-  compileUsingNoirWasm,
-  generateNoirContractInterface,
-  generateTypescriptContractInterface,
-} from './index.js';
-
-function isNargoAvailable() {
-  try {
-    execSync(`which nargo`);
-    return true;
-  } catch (error) {
-    return false;
-  }
-}
+import { compileUsingNoirWasm, generateNoirContractInterface, generateTypescriptContractInterface } from './index.js';
 
 describe('noir-compiler', () => {
   let projectPath: string;
@@ -30,18 +15,10 @@ describe('noir-compiler', () => {
     log = createDebugLogger('noir-compiler:test');
   });
 
-  const tests: Array<[string, (path: string, opts: { log: LogFn }) => Promise<ContractArtifact[]>]> = [
-    ['noir_wasm', compileUsingNoirWasm],
-  ];
-
-  if (isNargoAvailable()) {
-    tests.push(['nargo', compileUsingNargo]);
-  }
-
-  describe.each(tests)('using %s binary', (_, compileFn) => {
+  describe('using noir_wasm', () => {
     let compiled: ContractArtifact[];
     beforeAll(async () => {
-      compiled = await compileFn(projectPath, { log });
+      compiled = await compileUsingNoirWasm(projectPath, { log });
     });
 
     it('compiles the test contract', () => {
@@ -57,17 +34,5 @@ describe('noir-compiler', () => {
       const result = generateNoirContractInterface(compiled[0]);
       expect(result).toMatchSnapshot();
     });
-  });
-
-  const conditionalIt = isNargoAvailable() ? it : it.skip;
-  conditionalIt('both nargo and noir_wasm should compile identically', async () => {
-    const [noirWasmArtifact, nargoArtifact] = await Promise.all([
-      compileUsingNoirWasm(projectPath, { log }),
-      compileUsingNargo(projectPath, { log }),
-    ]);
-
-    const withoutDebug = ({ debug: _debug, ...rest }: ContractArtifact): Omit<ContractArtifact, 'debug'> => rest;
-
-    expect(nargoArtifact.map(withoutDebug)).toEqual(noirWasmArtifact.map(withoutDebug));
   });
 });
