+++ yarn-project/noir-compiler/src/index.ts
@@ -2,14 +2,11 @@ import { ContractArtifact } from '@aztec/foundation/abi';
 
 import { join } from 'path';
 
-import { CompileOpts, NargoContractCompiler } from './compile/nargo.js';
 import { OnDiskFileManager } from './compile/noir/file-manager/on-disk-file-manager.js';
 import { NoirWasmCompileOptions, NoirWasmContractCompiler } from './compile/noir/noir-wasm-compiler.js';
 import { generateContractArtifact } from './contract-interface-gen/abi.js';
-import NoirVersion from './noir-version.json' assert { type: 'json' };
 
-const { commit: NoirCommit, tag: NoirTag } = NoirVersion;
-export { NoirCommit, NoirTag };
+export { NoirVersion } from './noir-version.js';
 
 export { generateNoirContractInterface } from './contract-interface-gen/noir.js';
 export { generateTypescriptContractInterface } from './contract-interface-gen/typescript.js';
@@ -18,17 +15,6 @@ export { generateContractArtifact };
 /**
  * Compile Aztec.nr contracts in project path using a nargo binary available in the shell.
  * @param projectPath - Path to project.
- * @param opts - Compiler options.
- * @returns Compiled artifacts.
- */
-export async function compileUsingNargo(projectPath: string, opts: CompileOpts = {}): Promise<ContractArtifact[]> {
-  return (await new NargoContractCompiler(projectPath, opts).compile()).map(generateContractArtifact);
-}
-
-/**
- * Compile Aztec.nr contracts in project path using built-in noir_wasm.
- * @param projectPath - Path to project.
- * @param opts - Compiler options.
  * @returns Compiled artifacts.
  */
 export async function compileUsingNoirWasm(
