--- yarn-project/noir-compiler/src/noir-version.json
@@ -1,4 +1,0 @@
-{
-  "tag": "v0.17.0-aztec.5",
-  "commit": "b6819f6261e63cea3e653a1b874727a2b5919f5e"
-}
