+++ yarn-project/noir-compiler/tsconfig.json
@@ -10,5 +10,5 @@
       "path": "../foundation"
     }
   ],
-  "include": ["src", "src/*.json"]
+  "include": ["src"]
 }
