+++ yarn-project/noir-contracts/package.json
@@ -3,7 +3,6 @@
   "version": "0.1.0",
   "type": "module",
   "exports": {
-    ".": "./dest/index.js",
     "./artifacts": "./dest/artifacts/index.js",
     "./types": "./dest/types/index.js"
   },
