+++ yarn-project/noir-contracts/scripts/compile_all.sh
@@ -1,12 +1,7 @@
 #!/bin/bash
 # Compiles all noir contracts
 
-source ./scripts/nargo_check.sh
-
-echo "Checking noir version"
-nargo_check
-
 # Runs the compile scripts for all contracts.
 echo "Compiling all contracts"
 
-nargo compile --workspace --no-backend
+./scripts/compile.sh $(./scripts/get_all_contracts.sh)
